+++ app/javascript/data/doughnutchart/defaults.js
@@ -1,0 +1,82 @@
+import { sum } from '../../utils/array'
+
+let mapLabel = (chart, label, i) => {
+  let meta = chart.getDatasetMeta(0);
+  let ds = chart.data.datasets[0];
+  let arc = meta.data[i];
+  let custom = arc && arc.custom || {};
+  let getValueAtIndexOrDefault = Chart.helpers.getValueAtIndexOrDefault;
+  let arcOpts = chart.options.elements.arc;
+  let fill = custom.backgroundColor ? custom.backgroundColor : getValueAtIndexOrDefault(ds.backgroundColor, i, arcOpts.backgroundColor);
+  let stroke = custom.borderColor ? custom.borderColor : getValueAtIndexOrDefault(ds.borderColor, i, arcOpts.borderColor);
+  let bw = custom.borderWidth ? custom.borderWidth : getValueAtIndexOrDefault(ds.borderWidth, i, arcOpts.borderWidth);
+
+  return {
+    text: label,
+    fillStyle: fill,
+    strokeStyle: stroke,
+    lineWidth: bw,
+    hidden: isNaN(ds.data[i]) || meta.data[i].hidden,
+    index: i
+  }
+}
+
+const options = {
+  layout: {
+    padding: {
+      top: 15,
+    }
+  },
+  legend: {
+    position: "left",
+    labels: {
+      boxWidth: 12,
+      generateLabels: function(chart) {
+        let data = chart.data
+        return _.map(data.labels, mapLabel.bind(null, chart))
+      }
+    }
+  },
+  cutoutPercentage: 80,
+  plugins: {
+    datalabels: {
+      backgroundColor: function(context) {
+        return context.dataset.backgroundColor;
+      },
+      anchor: "end",
+      borderColor: 'white',
+      padding: function(context) {
+        var index = context.dataIndex;
+        var value = context.dataset.data[index];
+        return value < 10 ? { left: 6, right: 6, top: 3, bottom: 3 } : 5;
+      },
+      borderRadius: 100,
+      borderWidth: 1,
+      color: 'white',
+      display: true,
+      font: {
+        family: "'Helvetica Neue', 'Helvetica', 'Arial', sans-serif",
+      },
+      formatter: Math.round
+    }
+  }
+}
+
+export const doughnutData = function(data) {
+  let { colors, dataset } = data;
+  let [labels, values] = [_.keys(dataset), _.values(dataset)];
+
+  return  {
+    labels: labels,
+    total: sum(values),
+    datasets: [
+      { backgroundColor: colors, data: values }
+    ]
+  };
+}
+
+export const doughnutConfig = {
+  type: 'doughnut', 
+  plugins: [chartDataLabels],
+  options: options,
+}
